Moving an interfering context
