+++ quill-core/src/test/scala/io/getquill/UnlimitedOptionalEmbeddedSpec.scala
@@ -5,9 +5,10 @@ import io.getquill.ast.{ Entity, PropertyAlias }
 import io.getquill.base.Spec
 import io.getquill.context.mirror.Row
 import io.getquill.quat._
+import io.getquill.MirrorContexts.testContext
 
 class UnlimitedOptionalEmbeddedSpec extends Spec {
-  val ctx = testContext
+  val ctx = MirrorContexts.testContext
 
   import ctx._
