+++ quill-core/src/test/scala/io/getquill/context/BindMacroSpec.scala
@@ -2,8 +2,8 @@ package io.getquill.context
 
 import io.getquill.base.Spec
 import io.getquill.context.mirror.{ MirrorSession, Row }
-import io.getquill.testContext._
-import io.getquill.testContext
+import io.getquill.MirrorContexts.testContext._
+import io.getquill.MirrorContexts.testContext
 
 class BindMacroSpec extends Spec {
