+++ quill-core/src/test/scala/io/getquill/context/QueryMacroSpec.scala
@@ -2,8 +2,8 @@ package io.getquill.context
 
 import io.getquill.base.Spec
 import io.getquill.context.mirror.Row
-import io.getquill.testContext
-import io.getquill.testContext._
+import io.getquill.MirrorContexts.testContext
+import io.getquill.MirrorContexts.testContext._
 
 class QueryMacroSpec extends Spec {
