+++ quill-core/src/test/scala/io/getquill/norm/ApplyMapSpec.scala
@@ -1,13 +1,13 @@
 package io.getquill.norm
 
 import io.getquill.base.Spec
-import io.getquill.testContext.TestEntity
-import io.getquill.testContext.implicitOrd
-import io.getquill.testContext.qr1
-import io.getquill.testContext.qr2
-import io.getquill.testContext.query
-import io.getquill.testContext.quote
-import io.getquill.testContext.unquote
+import io.getquill.MirrorContexts.testContext.TestEntity
+import io.getquill.MirrorContexts.testContext.implicitOrd
+import io.getquill.MirrorContexts.testContext.qr1
+import io.getquill.MirrorContexts.testContext.qr2
+import io.getquill.MirrorContexts.testContext.query
+import io.getquill.MirrorContexts.testContext.quote
+import io.getquill.MirrorContexts.testContext.unquote
 import io.getquill.util.TraceConfig
 
 class ApplyMapSpec extends Spec {
