+++ quill-core/src/test/scala/io/getquill/norm/AttachToEntitySpec.scala
@@ -1,7 +1,7 @@
 package io.getquill.norm
 
 import io.getquill.ast.{ AscNullsFirst, Constant, Ident, Map, SortBy }
-import io.getquill.testContext._
+import io.getquill.MirrorContexts.testContext._
 import io.getquill.Query
 import io.getquill.base.Spec
