+++ quill-core/src/test/scala/io/getquill/norm/NormalizeNestedStructuresSpec.scala
@@ -1,11 +1,11 @@
 package io.getquill.norm
 
 import io.getquill.base.Spec
-import io.getquill.testContext.implicitOrd
-import io.getquill.testContext.qr1
-import io.getquill.testContext.qr2
-import io.getquill.testContext.quote
-import io.getquill.testContext.unquote
+import io.getquill.MirrorContexts.testContext.implicitOrd
+import io.getquill.MirrorContexts.testContext.qr1
+import io.getquill.MirrorContexts.testContext.qr2
+import io.getquill.MirrorContexts.testContext.quote
+import io.getquill.MirrorContexts.testContext.unquote
 
 class NormalizeNestedStructuresSpec extends Spec {
