+++ quill-core/src/test/scala/io/getquill/norm/NormalizeReturningSpec.scala
@@ -3,8 +3,9 @@ package io.getquill.norm
 import io.getquill.ReturnAction.{ ReturnColumns, ReturnRecord }
 import io.getquill.base.Spec
 import io.getquill.context.mirror.Row
-import io.getquill.{ MirrorIdiomReturningSingle, MirrorIdiomReturningMulti, testContext }
-import io.getquill.testContext._
+import io.getquill.{ MirrorIdiomReturningSingle, MirrorIdiomReturningMulti }
+import io.getquill.MirrorContexts.testContext._
+import io.getquill.MirrorContexts.testContext
 
 class NormalizeReturningSpec extends Spec {
