+++ quill-core/src/test/scala/io/getquill/norm/SymbolicReductionSpec.scala
@@ -1,11 +1,11 @@
 package io.getquill.norm
 
 import io.getquill.base.Spec
-import io.getquill.testContext.qr1
-import io.getquill.testContext.qr2
-import io.getquill.testContext.qr3
-import io.getquill.testContext.quote
-import io.getquill.testContext.unquote
+import io.getquill.MirrorContexts.testContext.qr1
+import io.getquill.MirrorContexts.testContext.qr2
+import io.getquill.MirrorContexts.testContext.qr3
+import io.getquill.MirrorContexts.testContext.quote
+import io.getquill.MirrorContexts.testContext.unquote
 import io.getquill.util.TraceConfig
 
 class SymbolicReductionSpec extends Spec { //hello
