+++ quill-core/src/test/scala/io/getquill/norm/capture/AvoidAliasConflictSpec.scala
@@ -1,6 +1,6 @@
 package io.getquill.norm.capture
 
-import io.getquill.testContext._
+import io.getquill.MirrorContexts.testContext._
 import io.getquill.Query
 import io.getquill.base.Spec
 import io.getquill.util.TraceConfig
