+++ quill-core/src/test/scala/io/getquill/quotation/FreeVariablesSpec.scala
@@ -2,11 +2,11 @@ package io.getquill.quotation
 
 import io.getquill.ast.IdentName
 import io.getquill.base.Spec
-import io.getquill.testContext.implicitOrd
-import io.getquill.testContext.qr1
-import io.getquill.testContext.qr2
-import io.getquill.testContext.quote
-import io.getquill.testContext.unquote
+import io.getquill.MirrorContexts.testContext.implicitOrd
+import io.getquill.MirrorContexts.testContext.qr1
+import io.getquill.MirrorContexts.testContext.qr2
+import io.getquill.MirrorContexts.testContext.quote
+import io.getquill.MirrorContexts.testContext.unquote
 
 class FreeVariablesSpec extends Spec {
