+++ quill-core/src/test/scala/io/getquill/quotation/IsDynamicSpec.scala
@@ -5,8 +5,8 @@ import io.getquill.ast.Property
 import io.getquill.ast.Renameable.Fixed
 import io.getquill.ast.Visibility.Visible
 import io.getquill.base.Spec
-import io.getquill.testContext.qr1
-import io.getquill.testContext.qrRegular
+import io.getquill.MirrorContexts.testContext.qr1
+import io.getquill.MirrorContexts.testContext.qrRegular
 
 class IsDynamicSpec extends Spec {
